Improve performances with zlib
Zlib is used if possible, otherwise the pure lua implementation of CRC32 is
used as previously.

As zlib is installed on most base linux installations, the dependency is
optional and the performance improvement can be dramatic (~30x faster when
testing zip.lua in isolation), I find it acceptable to add this to the project.